+++ NEWS.md
@@ -1,4 +1,9 @@
 
+# rqdatatable 0.1.3 2018/07/10
+
+  * Fix full join print glitch.
+  * data.table implementation of theta-join.
+
 # rqdatatable 0.1.2 2018/07/08
 
   * Adapt to instant execution path.